More cleanup in MatchResult
